refactor: update error render function
* feat: add custom title to render error

Update the `renderError` function to accept a
custom title so that it can be used for other
errors then just "Icon not found".

* feat: custom size error icons

Update the `renderError` function to accept a new parameter named
"iconSize" which specifies the width and height of the icon. Errors
rendered by the VueSimpleIcon will typically use `this.iconSize` for
this parameter.