Update the label used by Greenkeeper (#96)
* chore(greenkeeper): update the label used by Greenkeeper

...to "dependecies". As per (no official documentation as far as I know):
https://github.com/greenkeeperio/greenkeeper/issues/153#issuecomment-232732350

This brings it in line with the PRs generated by dependabot for security
fixes (e.g. #79). It is also just clearer in my opinion, especially for
people that don't know Greenkeeper.

* chore(greenkeeper): update label used by Greenkeeper from package.json

Tested the previous solution on one of my own repo's and it didn't work.
Let's hope this does work...